Extract rendering HtaccessResultTable to an external class
