+++ tests/HtaccessCommandTest.php
@@ -4,6 +4,7 @@ namespace Madewithlove;
 
 use Http\Adapter\Guzzle6\Client;
 use Http\Factory\Guzzle\ServerRequestFactory;
+use Madewithlove\Htaccess\TableRenderer;
 use Madewithlove\HtaccessClient;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Exception\RuntimeException;
@@ -27,7 +28,7 @@ final class HtaccessCommandTest extends TestCase
             new ServerRequestFactory()
         );
 
-        $this->command = new HtaccessCommand($htaccessClient);
+        $this->command = new HtaccessCommand($htaccessClient, new TableRenderer());
     }
 
     public function tearDown(): void
